refactor view definitions
