+++ pkg/gui/gui.go
@@ -223,30 +223,6 @@ type panelStates struct {
 	Merging    *MergingPanelState
 }
 
-type Views struct {
-	Status         *gocui.View
-	Files          *gocui.View
-	Branches       *gocui.View
-	RemoteBranches *gocui.View
-	Commits        *gocui.View
-	Stash          *gocui.View
-	Main           *gocui.View
-	Secondary      *gocui.View
-	Options        *gocui.View
-	Confirmation   *gocui.View
-	Menu           *gocui.View
-	CommitMessage  *gocui.View
-	CommitFiles    *gocui.View
-	SubCommits     *gocui.View
-	Information    *gocui.View
-	AppStatus      *gocui.View
-	Search         *gocui.View
-	SearchPrefix   *gocui.View
-	Limit          *gocui.View
-	Suggestions    *gocui.View
-	Extras         *gocui.View
-}
-
 type searchingState struct {
 	view         *gocui.View
 	isSearching  bool
@@ -594,121 +570,6 @@ func (gui *Gui) Run(filterPath string) error {
 	gui.c.Log.Info("starting main loop")
 
 	return gui.g.MainLoop()
-}
-
-func (gui *Gui) createAllViews() error {
-	viewNameMappings := []struct {
-		viewPtr **gocui.View
-		name    string
-	}{
-		{viewPtr: &gui.Views.Status, name: "status"},
-		{viewPtr: &gui.Views.Files, name: "files"},
-		{viewPtr: &gui.Views.Branches, name: "branches"},
-		{viewPtr: &gui.Views.RemoteBranches, name: "remoteBranches"},
-		{viewPtr: &gui.Views.Commits, name: "commits"},
-		{viewPtr: &gui.Views.Stash, name: "stash"},
-		{viewPtr: &gui.Views.CommitFiles, name: "commitFiles"},
-		{viewPtr: &gui.Views.SubCommits, name: "subCommits"},
-		{viewPtr: &gui.Views.Main, name: "main"},
-		{viewPtr: &gui.Views.Secondary, name: "secondary"},
-		{viewPtr: &gui.Views.Options, name: "options"},
-		{viewPtr: &gui.Views.AppStatus, name: "appStatus"},
-		{viewPtr: &gui.Views.Information, name: "information"},
-		{viewPtr: &gui.Views.Search, name: "search"},
-		{viewPtr: &gui.Views.SearchPrefix, name: "searchPrefix"},
-		{viewPtr: &gui.Views.CommitMessage, name: "commitMessage"},
-		{viewPtr: &gui.Views.Menu, name: "menu"},
-		{viewPtr: &gui.Views.Suggestions, name: "suggestions"},
-		{viewPtr: &gui.Views.Confirmation, name: "confirmation"},
-		{viewPtr: &gui.Views.Limit, name: "limit"},
-		{viewPtr: &gui.Views.Extras, name: "extras"},
-	}
-
-	var err error
-	for _, mapping := range viewNameMappings {
-		*mapping.viewPtr, err = gui.prepareView(mapping.name)
-		if err != nil && err.Error() != UNKNOWN_VIEW_ERROR_MSG {
-			return err
-		}
-	}
-
-	gui.Views.Options.Frame = false
-	gui.Views.Options.FgColor = theme.OptionsColor
-
-	gui.Views.SearchPrefix.BgColor = gocui.ColorDefault
-	gui.Views.SearchPrefix.FgColor = gocui.ColorGreen
-	gui.Views.SearchPrefix.Frame = false
-	gui.setViewContent(gui.Views.SearchPrefix, SEARCH_PREFIX)
-
-	gui.Views.Stash.Title = gui.c.Tr.StashTitle
-	gui.Views.Stash.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.Commits.Title = gui.c.Tr.CommitsTitle
-	gui.Views.Commits.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.CommitFiles.Title = gui.c.Tr.CommitFiles
-	gui.Views.CommitFiles.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.SubCommits.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.Branches.Title = gui.c.Tr.BranchesTitle
-	gui.Views.Branches.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.RemoteBranches.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.Files.Title = gui.c.Tr.FilesTitle
-	gui.Views.Files.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.Secondary.Title = gui.c.Tr.DiffTitle
-	gui.Views.Secondary.Wrap = true
-	gui.Views.Secondary.FgColor = theme.GocuiDefaultTextColor
-	gui.Views.Secondary.IgnoreCarriageReturns = true
-	gui.Views.Secondary.CanScrollPastBottom = gui.c.UserConfig.Gui.ScrollPastBottom
-
-	gui.Views.Main.Title = gui.c.Tr.DiffTitle
-	gui.Views.Main.Wrap = true
-	gui.Views.Main.FgColor = theme.GocuiDefaultTextColor
-	gui.Views.Main.IgnoreCarriageReturns = true
-	gui.Views.Main.CanScrollPastBottom = gui.c.UserConfig.Gui.ScrollPastBottom
-
-	gui.Views.Limit.Title = gui.c.Tr.NotEnoughSpace
-	gui.Views.Limit.Wrap = true
-
-	gui.Views.Status.Title = gui.c.Tr.StatusTitle
-	gui.Views.Status.FgColor = theme.GocuiDefaultTextColor
-
-	gui.Views.Search.BgColor = gocui.ColorDefault
-	gui.Views.Search.FgColor = gocui.ColorGreen
-	gui.Views.Search.Frame = false
-	gui.Views.Search.Editable = true
-
-	gui.Views.AppStatus.BgColor = gocui.ColorDefault
-	gui.Views.AppStatus.FgColor = gocui.ColorCyan
-	gui.Views.AppStatus.Frame = false
-	gui.Views.AppStatus.Visible = false
-
-	gui.Views.CommitMessage.Visible = false
-	gui.Views.CommitMessage.Title = gui.c.Tr.CommitMessage
-	gui.Views.CommitMessage.FgColor = theme.GocuiDefaultTextColor
-	gui.Views.CommitMessage.Editable = true
-	gui.Views.CommitMessage.Editor = gocui.EditorFunc(gui.commitMessageEditor)
-
-	gui.Views.Confirmation.Visible = false
-
-	gui.Views.Suggestions.Visible = false
-
-	gui.Views.Menu.Visible = false
-
-	gui.Views.Information.BgColor = gocui.ColorDefault
-	gui.Views.Information.FgColor = gocui.ColorGreen
-	gui.Views.Information.Frame = false
-
-	gui.Views.Extras.Title = gui.c.Tr.CommandLog
-	gui.Views.Extras.FgColor = theme.GocuiDefaultTextColor
-	gui.Views.Extras.Autoscroll = true
-	gui.Views.Extras.Wrap = true
-
-	return nil
 }
 
 func (gui *Gui) RunAndHandleError(filterPath string) error {
